Fix incorrect return type in smbase/crc.h
Also cast the type back to unsigned long in datablok.cpp to keep
format from complaining. uint32_t should never be longer than
unsigned long

+++ src/smbase/datablok.cpp
@@ -250,7 +250,7 @@ void DataBlock::print(char const *label, int bytesPerLine) const
   if (label) {
     printf("---- %s, length = %d, crc32 = 0x%lX ---- {\n",
            label, getDataLen(),
-           crc32(getDataC(), getDataLen()));
+           (long unsigned int)crc32(getDataC(), getDataLen()));
   }
 
   int cursor = 0;
